[feat] 키보드 숨기는 부분 util로 빼고, mypage에서도 사용할 수 있도록 구현
