+++ iOS/Macro/Macro/Scene/Write/InterfaceAdapters/View/WriteViewController.swift
@@ -320,15 +320,3 @@ extension WriteViewController: UITextFieldDelegate, UITextViewDelegate {
     }
 }
 
-// 키보드 숨기기
-extension UIViewController {
-    func hideKeyboardWhenTappedAround() {
-        let tap = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
-        tap.cancelsTouchesInView = false
-        view.addGestureRecognizer(tap)
-    }
-    
-    @objc func dismissKeyboard() {
-        view.endEditing(true)
-    }
-}
