+++ iOS/Macro/Macro/Util/UIViewController+.swift
@@ -1,0 +1,21 @@
+//
+//  UIViewController+.swift
+//  Macro
+//
+//  Created by 김나훈 on 12/1/23.
+//
+
+import UIKit
+
+// 키보드 숨기기
+extension UIViewController {
+    func hideKeyboardWhenTappedAround() {
+        let tap = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        tap.cancelsTouchesInView = false
+        view.addGestureRecognizer(tap)
+    }
+    
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+}
